12-12-2024/proyecto_api/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-allauth==65.1.0
django-filter==24.3
django-model-utils==5.0.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
Markdown==3.7
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
requests==2.32.3